Etc: 코드 정리
- 불필요한 파일 삭제
- ProfileEdit ViewHolder 파일 분리
- 더 명확한 파라미터 네이밍으로 수정
