+++ Pyonsnal-Color/Pyonsnal-Color/ProfileEdit/ProfileEditInteractor.swift
@@ -34,8 +34,8 @@ final class ProfileEditInteractor: PresentableInteractor<ProfileEditPresentable>
     
     var isEditButtonEnabled: AnyPublisher<Bool, Never> {
         Publishers.CombineLatest(isNicknameValid(), isChangedProfileImage())
-            .map { nicknameValid, profileEdited in
-                if nicknameValid || profileEdited || (nicknameValid && profileEdited) {
+            .map { nicknameValid, profileImageEdited in
+                if nicknameValid || profileImageEdited || (nicknameValid && profileImageEdited) {
                     return true
                 }
                 return false
